Move tests that require Java 11 into their own file
